Fix exporting environment variables with whitespaces
Previously, when attemping to export enviornment variables with spaces
(e.g. NODE_OPTIONS="-r [...]"), we would get an error message: `export: not valid in this context: -r`.
This is because xargs doesn't properly support strings with spaces.

To fix this, we are doing two things:
1. Switch to doing an `eval` of the script's output (which exports the
   env vars)
2. Properly escape the environment variables.

This results in us being able to export environment variables with
spaces.